Added doc for correl

+++ crappy2/blocks/_correl.py
@@ -59,7 +59,7 @@ class Correl(MasterBlock):
       t2 = time()-1
       for i in range(2): # Drop the first images
         self.inputs[0].recv()
-      correl.setOrig(self.inputs[0].recv()) # This is the only time the original picture is set, so the residual may increase if lightning vary or large displacements are reached
+      correl.setOrig(self.inputs[0].recv().astype(np.float32)) # This is the only time the original picture is set, so the residual may increase if lightning vary or large displacements are reached
       correl.prepare()
       while True:
         t1 = time()
